requirements.txt
matplotlib>=3.2.2
numpy>=1.19.5
opencv-python>=4.1.2
Pillow>=7.1.2
torch>=1.10.0
torchvision>=0.11.1
albumentations == 1.1.0